refactor: move status from operations to commands package
Status checking is a read-only query about pack state, not an operation
that modifies files. This change:

- Moves all status types and implementation to pkg/packs/commands/status.go
- Removes status files from pkg/packs/operations/
- Updates all imports to use commands.StatusXXX instead of operations.StatusXXX
- Fixes the circular dependency issue by allowing status to import handlers directly
- Implements proper status checking for both configuration and code execution handlers

The status implementation now properly checks:
- Configuration handlers: Checks if intermediate symlinks exist
- Code execution handlers: Checks sentinels in the datastore
