+++ pkg/packs/commands/adopt.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/arthur-debert/dodot/pkg/filesystem"
 	"github.com/arthur-debert/dodot/pkg/logging"
-	"github.com/arthur-debert/dodot/pkg/packs/operations"
 	"github.com/arthur-debert/dodot/pkg/packs/execution"
 	"github.com/arthur-debert/dodot/pkg/paths"
 	"github.com/arthur-debert/dodot/pkg/types"
@@ -123,17 +122,17 @@ func (c *AdoptCommand) ExecuteForPack(pack types.Pack, opts execution.Options) (
 	}
 
 	// Create status result
-	statusResult := &operations.StatusResult{
+	statusResult := &StatusResult{
 		Name:   pack.Name,
 		Status: "success",
-		Files:  []operations.FileStatus{},
+		Files:  []FileStatus{},
 	}
 
 	// Add adopted files to status
 	for _, fileName := range adoptedFiles {
-		statusResult.Files = append(statusResult.Files, operations.FileStatus{
+		statusResult.Files = append(statusResult.Files, FileStatus{
 			Path:   fileName,
-			Status: operations.Status{State: operations.StatusStateSuccess},
+			Status: Status{State: StatusStateSuccess},
 		})
 	}
