+++ pkg/packs/commands/status.go
@@ -1,13 +1,21 @@
 package commands
 
 import (
+	"fmt"
+	"path/filepath"
+	"time"
+
 	"github.com/arthur-debert/dodot/pkg/datastore"
 	"github.com/arthur-debert/dodot/pkg/filesystem"
+	"github.com/arthur-debert/dodot/pkg/handlers"
 	"github.com/arthur-debert/dodot/pkg/logging"
+	"github.com/arthur-debert/dodot/pkg/operations"
+	"github.com/arthur-debert/dodot/pkg/packs/discovery"
 	"github.com/arthur-debert/dodot/pkg/packs/execution"
-	"github.com/arthur-debert/dodot/pkg/packs/operations"
 	"github.com/arthur-debert/dodot/pkg/paths"
+	"github.com/arthur-debert/dodot/pkg/rules"
 	"github.com/arthur-debert/dodot/pkg/types"
+	"github.com/arthur-debert/dodot/pkg/ui/display"
 )
 
 // StatusCommand implements the "status" command using the pack execution.
@@ -16,6 +24,59 @@ type StatusCommand struct{}
 // Name returns the command name.
 func (c *StatusCommand) Name() string {
 	return "status"
+}
+
+// StatusOptions contains options for getting pack status
+type StatusOptions struct {
+	Pack       types.Pack
+	DataStore  datastore.DataStore
+	FileSystem types.FS
+	Paths      paths.Paths
+}
+
+// StatusState represents the state of a deployment
+type StatusState string
+
+const (
+	// StatusStatePending indicates the action has not been executed yet
+	StatusStatePending StatusState = "pending"
+	// StatusStateReady indicates the action was executed and is ready
+	StatusStateReady StatusState = "ready"
+	// StatusStateSuccess indicates success
+	StatusStateSuccess StatusState = "success"
+	// StatusStateMissing indicates missing files
+	StatusStateMissing StatusState = "missing"
+	// StatusStateError indicates an error occurred
+	StatusStateError StatusState = "error"
+	// StatusStateIgnored indicates the file/pack is ignored
+	StatusStateIgnored StatusState = "ignored"
+	// StatusStateConfig indicates this is a config file
+	StatusStateConfig StatusState = "config"
+)
+
+// Status represents the status of a single item
+type Status struct {
+	State     StatusState
+	Message   string
+	Timestamp *time.Time
+}
+
+// FileStatus represents the status of a single file
+type FileStatus struct {
+	Handler        string
+	Path           string
+	Status         Status
+	AdditionalInfo string
+}
+
+// StatusResult represents the complete status of a pack
+type StatusResult struct {
+	Name      string
+	Path      string
+	HasConfig bool
+	IsIgnored bool
+	Status    string
+	Files     []FileStatus
 }
 
 // ExecuteForPack executes the "status" command for a single pack.
@@ -44,15 +105,15 @@ func (c *StatusCommand) ExecuteForPack(pack types.Pack, opts execution.Options) 
 	// Create datastore for status checking
 	dataStore := datastore.New(fs, pathsInstance)
 
-	// Get pack status using the centralized GetStatus function
-	statusOpts := operations.StatusOptions{
+	// Get pack status
+	statusOpts := StatusOptions{
 		Pack:       pack,
 		DataStore:  dataStore,
 		FileSystem: fs,
 		Paths:      pathsInstance,
 	}
 
-	packStatus, err := operations.GetStatus(statusOpts)
+	packStatus, err := GetStatus(statusOpts)
 	if err != nil {
 		logger.Error().
 			Err(err).
@@ -77,4 +138,342 @@ func (c *StatusCommand) ExecuteForPack(pack types.Pack, opts execution.Options) 
 		Error:                 nil,
 		CommandSpecificResult: packStatus,
 	}, nil
+}
+
+// GetStatus retrieves the status of a pack
+func GetStatus(opts StatusOptions) (*StatusResult, error) {
+	logger := logging.GetLogger("commands.status")
+	logger.Debug().
+		Str("pack", opts.Pack.Name).
+		Msg("Getting pack status")
+
+	result := &StatusResult{
+		Name:      opts.Pack.Name,
+		Path:      opts.Pack.Path,
+		HasConfig: false, // TODO: Check for .dodot.toml file
+		IsIgnored: false, // TODO: Check for .dodotignore file
+		Status:    "unknown",
+		Files:     []FileStatus{},
+	}
+
+	// If pack is ignored, return early
+	if result.IsIgnored {
+		result.Status = "ignored"
+		return result, nil
+	}
+
+	// Get matches for this pack
+	matches, err := rules.NewMatcher().GetMatchesFS([]types.Pack{opts.Pack}, opts.FileSystem)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get matches: %w", err)
+	}
+
+	// Check status for each match
+	for _, match := range matches {
+		fileStatus, err := getHandlerStatus(match, opts.Pack, opts.DataStore, opts.FileSystem, opts.Paths)
+		if err != nil {
+			logger.Error().
+				Err(err).
+				Str("file", match.Path).
+				Str("handler", match.HandlerName).
+				Msg("Failed to get handler status")
+			continue
+		}
+
+		result.Files = append(result.Files, FileStatus{
+			Handler:        match.HandlerName,
+			Path:           match.Path,
+			Status:         fileStatus,
+			AdditionalInfo: "",
+		})
+	}
+
+	// Determine overall pack status based on file statuses
+	result.Status = determinePackStatus(result.Files)
+
+	return result, nil
+}
+
+// getHandlerStatus checks the deployment status for a specific match
+func getHandlerStatus(match rules.RuleMatch, pack types.Pack, dataStore datastore.DataStore, fs types.FS, pathsInstance paths.Paths) (Status, error) {
+	logger := logging.GetLogger("commands.status.handler")
+
+	// Get handler category to determine how to check status
+	category := handlers.HandlerRegistry.GetHandlerCategory(match.HandlerName)
+
+	switch category {
+	case operations.CategoryConfiguration:
+		// For configuration handlers, check if intermediate links exist
+		linkPath := pathsInstance.PackHandlerDir(pack.Name, match.HandlerName)
+		targetPath := filepath.Join(linkPath, match.Path)
+
+		_, err := fs.Stat(targetPath)
+		if err == nil {
+			return Status{
+				State:   StatusStateReady,
+				Message: "Linked",
+			}, nil
+		}
+
+		return Status{
+			State:   StatusStatePending,
+			Message: "Not linked",
+		}, nil
+
+	case operations.CategoryCodeExecution:
+		// For code execution handlers, check sentinels in datastore
+		// Generate sentinel based on handler type and match
+		sentinel := generateSentinel(match)
+
+		exists, err := dataStore.HasSentinel(pack.Name, match.HandlerName, sentinel)
+		if err != nil {
+			logger.Error().
+				Err(err).
+				Str("pack", pack.Name).
+				Str("handler", match.HandlerName).
+				Str("sentinel", sentinel).
+				Msg("Failed to check sentinel")
+			return Status{
+				State:   StatusStateError,
+				Message: fmt.Sprintf("Failed to check status: %v", err),
+			}, nil
+		}
+
+		if exists {
+			return Status{
+				State:   StatusStateReady,
+				Message: "Installed",
+			}, nil
+		}
+
+		return Status{
+			State:   StatusStatePending,
+			Message: "Not installed",
+		}, nil
+
+	default:
+		return Status{
+			State:   StatusStateError,
+			Message: fmt.Sprintf("Unknown handler category for %s", match.HandlerName),
+		}, nil
+	}
+}
+
+// generateSentinel generates a sentinel key for a code execution handler match
+func generateSentinel(match rules.RuleMatch) string {
+	// For homebrew, use the Brewfile path
+	if match.HandlerName == "homebrew" {
+		return match.Path
+	}
+
+	// For install scripts, use the script path
+	if match.HandlerName == "install" {
+		return match.Path
+	}
+
+	// Default: use the file path
+	return match.Path
+}
+
+// determinePackStatus calculates the overall pack status from file statuses
+func determinePackStatus(files []FileStatus) string {
+	if len(files) == 0 {
+		return "empty"
+	}
+
+	hasError := false
+	hasSuccess := false
+	hasPending := false
+
+	for _, file := range files {
+		switch file.Status.State {
+		case StatusStateError:
+			hasError = true
+		case StatusStateReady, StatusStateSuccess:
+			hasSuccess = true
+		case StatusStatePending, StatusStateMissing:
+			hasPending = true
+		}
+	}
+
+	if hasError {
+		return "error"
+	}
+	if hasPending && !hasSuccess {
+		return "pending"
+	}
+	if hasPending && hasSuccess {
+		return "partial"
+	}
+	if hasSuccess {
+		return "success"
+	}
+
+	return "unknown"
+}
+
+// StatusCommandOptions contains options for the status command
+type StatusCommandOptions struct {
+	// DotfilesRoot is the root directory containing packs
+	DotfilesRoot string
+
+	// PackNames specifies which packs to check status for
+	// If empty, all packs are checked
+	PackNames []string
+
+	// Paths provides system paths (optional, will be created if not provided)
+	Paths types.Pather
+
+	// FileSystem to use (optional, defaults to OS filesystem)
+	FileSystem types.FS
+}
+
+// GetPacksStatus shows the deployment status of specified packs
+// This is a query operation that uses core pack discovery but doesn't execute handlers.
+func GetPacksStatus(opts StatusCommandOptions) (*display.PackCommandResult, error) {
+	logger := logging.GetLogger("pack.status")
+	logger.Debug().
+		Str("dotfilesRoot", opts.DotfilesRoot).
+		Strs("packNames", opts.PackNames).
+		Msg("Starting status command")
+
+	// Track any errors encountered
+	var errors []error
+
+	// Initialize filesystem if not provided
+	if opts.FileSystem == nil {
+		opts.FileSystem = filesystem.NewOS()
+	}
+
+	// Initialize paths if not provided
+	if opts.Paths == nil {
+		p, err := paths.New(opts.DotfilesRoot)
+		if err != nil {
+			return nil, fmt.Errorf("failed to initialize paths: %w", err)
+		}
+		opts.Paths = p
+	}
+
+	// Use core pack discovery (consistent with on/off commands)
+	selectedPacks, err := discovery.DiscoverAndSelectPacksFS(opts.DotfilesRoot, opts.PackNames, opts.FileSystem)
+	if err != nil {
+		logger.Error().Err(err).Msg("Failed to discover packs")
+		return nil, err
+	}
+
+	logger.Info().
+		Int("packCount", len(selectedPacks)).
+		Msg("Found packs to check")
+
+	// Create datastore for status checking
+	dataStore := datastore.New(opts.FileSystem, opts.Paths.(paths.Paths))
+
+	// Build command result
+	result := &display.PackCommandResult{
+		Command:   "status",
+		DryRun:    false, // Status is always a query, never a dry run
+		Timestamp: time.Now(),
+		Packs:     make([]display.DisplayPack, 0, len(selectedPacks)),
+		// Status command doesn't have a message
+		Message: "",
+	}
+
+	// Process each pack using centralized status logic
+	for _, p := range selectedPacks {
+		// Get pack status using the centralized GetStatus function
+		statusOpts := StatusOptions{
+			Pack:       p,
+			DataStore:  dataStore,
+			FileSystem: opts.FileSystem,
+			Paths:      opts.Paths.(paths.Paths),
+		}
+
+		packStatus, err := GetStatus(statusOpts)
+		if err != nil {
+			logger.Error().
+				Err(err).
+				Str("pack", p.Name).
+				Msg("Failed to get pack status")
+			errors = append(errors, fmt.Errorf("pack %s: status check failed: %w", p.Name, err))
+			// Continue with other packs even if one fails
+			continue
+		}
+
+		// Convert to display format using existing conversion logic
+		displayPack := convertStatusToDisplayPack(packStatus)
+		result.Packs = append(result.Packs, displayPack)
+	}
+
+	logger.Info().
+		Int("packsProcessed", len(result.Packs)).
+		Int("errors", len(errors)).
+		Msg("Status command completed")
+
+	// Return error if any packs failed (but still return partial results)
+	if len(errors) > 0 {
+		return result, fmt.Errorf("status command encountered %d errors", len(errors))
+	}
+
+	return result, nil
+}
+
+// convertStatusToDisplayPack converts StatusResult to display.DisplayPack
+func convertStatusToDisplayPack(status *StatusResult) display.DisplayPack {
+	displayPack := display.DisplayPack{
+		Name:      status.Name,
+		HasConfig: status.HasConfig,
+		IsIgnored: status.IsIgnored,
+		Status:    status.Status,
+		Files:     make([]display.DisplayFile, 0, len(status.Files)),
+	}
+
+	// Convert each file status
+	for _, file := range status.Files {
+		displayFile := display.DisplayFile{
+			Handler:        file.Handler,
+			Path:           file.Path,
+			Status:         statusStateToDisplayStatus(file.Status.State),
+			Message:        file.Status.Message,
+			LastExecuted:   file.Status.Timestamp,
+			HandlerSymbol:  display.GetHandlerSymbol(file.Handler),
+			AdditionalInfo: file.AdditionalInfo,
+		}
+		displayPack.Files = append(displayPack.Files, displayFile)
+	}
+
+	// Add special files if present
+	if status.IsIgnored {
+		displayPack.Files = append([]display.DisplayFile{{
+			Path:   ".dodotignore",
+			Status: "ignored",
+		}}, displayPack.Files...)
+	}
+	if status.HasConfig {
+		displayPack.Files = append([]display.DisplayFile{{
+			Path:   ".dodot.toml",
+			Status: "config",
+		}}, displayPack.Files...)
+	}
+
+	return displayPack
+}
+
+// statusStateToDisplayStatus converts internal status states to display status strings
+func statusStateToDisplayStatus(state StatusState) string {
+	switch state {
+	case StatusStateReady, StatusStateSuccess:
+		return "success"
+	case StatusStateMissing:
+		return "queue"
+	case StatusStatePending:
+		return "queue"
+	case StatusStateError:
+		return "error"
+	case StatusStateIgnored:
+		return "ignored"
+	case StatusStateConfig:
+		return "config"
+	default:
+		return "unknown"
+	}
 }
